Increase monit timeout on Postgres job start
We've discovered a bug in the following scenario:

Using HA Postgres, with one node down, and the other attempting to
start.

The node attempting to start will take about 30-40 seconds to bootstrap,
waiting for the other node to potentially come online. By default,
`monit start` timeout is 30 seconds, which can in some cases cause a
loop where Postgres attempts to start and gets killed by monit in a
loop. This commit fixes that by extending the timeout to 60 seconds.